corregi errores de tipo de datos

+++ kernel/globals/globals.go
@@ -7,16 +7,16 @@ import (
 )
 
 type Config struct {
-	Ip_memory               string `json:"ip_memory"`
-	Port_memory             int    `json:"port_memory"`
-	Ip_Kernel               string `json:"ip_Kernel"`
-	Port_kernel             int    `json:"port_kernel"`
-	Scheduler_algorithm     string `json:"scheduler_algorithm"`
-	Ready_ingress_algorithm string `json:"Ready_ingress_algorithm"`
-	Alpha                   int    `json:"alpha"`
-	Initial_estimate        int    `json:"initial_estimate"`
-	Suspension_time         int    `json:"suspension_time"`
-	Log_level               string `json:"log_level"`
+	Ip_memory               string  `json:"ip_memory"`
+	Port_memory             int     `json:"port_memory"`
+	Ip_Kernel               string  `json:"ip_Kernel"`
+	Port_kernel             int     `json:"port_kernel"`
+	Scheduler_algorithm     string  `json:"scheduler_algorithm"`
+	Ready_ingress_algorithm string  `json:"Ready_ingress_algorithm"`
+	Alpha                   float32 `json:"alpha"`
+	Initial_estimate        float32 `json:"initial_estimate"`
+	Suspension_time         int     `json:"suspension_time"`
+	Log_level               string  `json:"log_level"`
 }
 
 var ClientConfig *Config
